v.2.0.0.17 rewrite wolf.system with c

+++ engine/src/wolf.system/base64/quicktimebase64.c
@@ -113,7 +113,7 @@ int quicktime_base64_decode_len(const char *bufcoded) {
   while (pr2six[*(bufin++)] <= 63)
     ;
 
-  nprbytes = (bufin - (const unsigned char *)bufcoded) - 1;
+  nprbytes = (int)((bufin - (const unsigned char *)bufcoded) - 1);
   nbytesdecoded = ((nprbytes + 3) / 4) * 3;
 
   return nbytesdecoded + 1;
@@ -128,7 +128,7 @@ int quicktime_base64_decode(char *bufplain, const char *bufcoded) {
   bufin = (const unsigned char *)bufcoded;
   while (pr2six[*(bufin++)] <= 63)
     ;
-  nprbytes = (bufin - (const unsigned char *)bufcoded) - 1;
+  nprbytes = (int)((bufin - (const unsigned char *)bufcoded) - 1);
   nbytesdecoded = ((nprbytes + 3) / 4) * 3;
 
   bufout = (unsigned char *)bufplain;
@@ -192,5 +192,5 @@ int quicktime_base64_encode(char *encoded, const char *string, int len) {
   }
 
   *p++ = '\0';
-  return p - encoded;
+  return (int)(p - encoded);
 }
